+++ MVC_Project.Domain/Entities/IEntity.cs
@@ -5,6 +5,6 @@ using System.Text;
 namespace MVC_Project.Domain.Entities {
 
     public interface IEntity {
-        int id { get; set; }
+        Int64 id { get; set; }
     }
 }
