+++ MVC_Project.Domain/Entities/Role.cs
@@ -5,7 +5,7 @@ using System.Text;
 namespace MVC_Project.Domain.Entities {
 
     public class Role : IEntity {
-        public virtual int id { get; set; }
+        public virtual Int64 id { get; set; }
         public virtual Guid uuid { get; set; }
         public virtual string name { get; set; }
         public virtual string code { get; set; }
