+++ MVC_Project.Domain/Entities/RolePermission.cs
@@ -8,7 +8,7 @@ namespace MVC_Project.Domain.Entities
 {
     public class RolePermission : IEntity
     {
-        public virtual int id { get; set; }
+        public virtual Int64 id { get; set; }
         public virtual Role role { get; set; }
         public virtual Permission permission { get; set; }
         public virtual Account account { get; set; }
